refactor : LottoReward 제거 Rank에 reward추가
