+++ src/test/java/lotto/domain/RankTest.java
@@ -1,5 +1,6 @@
 package lotto.domain;
 
+import lotto.domain.Money.ImmutableMoney;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
@@ -27,5 +28,24 @@ public class RankTest {
     @MethodSource("provideForRank")
     void rank(LottoMatchCount matchCount, boolean containBonusNumber, Rank rank) {
         assertThat(Rank.rankValue(matchCount, containBonusNumber)).isEqualTo(rank);
+    }
+
+    private static Stream<Arguments> provideLottoReward() {
+        return Stream.of(
+                Arguments.of(new ImmutableMoney(2_000_000_000), Rank.FIRST),
+                Arguments.of(new ImmutableMoney(30_000_000), Rank.SECOND),
+                Arguments.of(new ImmutableMoney(1_500_000), Rank.THIRD),
+                Arguments.of(new ImmutableMoney(50_000), Rank.FOURTH),
+                Arguments.of(new ImmutableMoney(5_000), Rank.FIFTH),
+                Arguments.of(new ImmutableMoney(0), Rank.SIXTH),
+                Arguments.of(new ImmutableMoney(0), Rank.SEVENTH),
+                Arguments.of(new ImmutableMoney(0), Rank.EIGHTH)
+        );
+    }
+
+    @ParameterizedTest(name = "등수에 따른 상금을 반환한다.")
+    @MethodSource(value = "provideLottoReward")
+    void reward(ImmutableMoney reward, Rank rank) {
+        assertThat(rank.reward()).isEqualTo(reward);
     }
 }
